refactor: centralize enums
