+++ LYBT.Module.Users/Dtos/UserDto.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using LYBT.Common.Enums;
+using LYBT.Common.Enums.Users;
 
 namespace LYBT.Module.Users.Dtos {
     /// <summary>
